+++ projects/swimlane/ngx-ui/src/lib/components/input/input-prefix.component.ts
@@ -3,7 +3,9 @@ import { Component, ChangeDetectionStrategy } from '@angular/core';
 @Component({
   exportAs: 'ngxInputPrefix',
   selector: 'ngx-input-prefix',
-  template: `<ng-content></ng-content>`,
-  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: `
+    <ng-content></ng-content>
+  `,
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class InputPrefixComponent {}