+++ projects/swimlane/ngx-ui/src/lib/components/section/section.component.ts
@@ -29,7 +29,7 @@ export class SectionComponent {
 
   @Output() toggle = new EventEmitter();
 
-  @ContentChild(SectionHeaderComponent, { static: false }) headerComp: SectionHeaderComponent;
+  @ContentChild(SectionHeaderComponent) headerComp: SectionHeaderComponent;
 
   onSectionClicked(): void {
     this.sectionCollapsed = !this.sectionCollapsed;