+++ projects/swimlane/ngx-ui/src/lib/components/select/contains-filter.util.spec.ts
@@ -2,7 +2,7 @@ import { containsFilter } from './contains-filter.util';
 
 describe('containsFilter', () => {
   it('should be true when not case sensitive and filter exists', () => {
-    const res = containsFilter('test', 'test', { });
+    const res = containsFilter('test', 'test', {});
     expect(res).toBeTruthy();
   });
 
@@ -12,21 +12,25 @@ describe('containsFilter', () => {
   });
 
   it('should be false of !value', () => {
-    const res = containsFilter(undefined, undefined, { });
+    const res = containsFilter(undefined, undefined, {});
     expect(res).toBeFalsy();
   });
 
   it('should traverse object value', () => {
-    const res = containsFilter({
-      test: 'test',
-      nottest: '1'
-    }, '1', { });
+    const res = containsFilter(
+      {
+        test: 'test',
+        nottest: '1'
+      },
+      '1',
+      {}
+    );
 
     expect(res).toBeTruthy();
   });
 
   it('should be undefined if value is not string or object', () => {
-    const res = containsFilter(1, '1', { });
+    const res = containsFilter(1, '1', {});
     expect(res).toBeUndefined();
   });
 });