+++ projects/swimlane/ngx-ui/src/lib/components/select/select.component.ts
@@ -70,85 +70,113 @@ export class SelectComponent implements ControlValueAccessor, OnDestroy {
   @Input() identifier: string;
 
   @Input()
-  get minSelections() { return this._minSelections; }
+  get minSelections() {
+    return this._minSelections;
+  }
   set minSelections(minSelections) {
     this._minSelections = coerceNumberProperty(minSelections);
   }
 
   @Input()
-  get maxSelections() { return this._maxSelections; };
+  get maxSelections() {
+    return this._maxSelections;
+  }
   set maxSelections(maxSelections) {
     this._maxSelections = coerceNumberProperty(maxSelections);
   }
 
   @Input()
-  get autofocus() { return this._autofocus; };
+  get autofocus() {
+    return this._autofocus;
+  }
   set autofocus(autofocus) {
     this._autofocus = coerceBooleanProperty(autofocus);
   }
 
   @Input()
-  get allowClear() { return this._allowClear; }
+  get allowClear() {
+    return this._allowClear;
+  }
   set allowClear(allowClear) {
     this._allowClear = coerceBooleanProperty(allowClear);
   }
 
   @Input()
-  get allowAdditions() { return this._allowAdditions; }
+  get allowAdditions() {
+    return this._allowAdditions;
+  }
   set allowAdditions(allowAdditions) {
     this._allowAdditions = coerceBooleanProperty(allowAdditions);
   }
 
   @Input()
-  get disableDropdown() { return this._disableDropdown; }
+  get disableDropdown() {
+    return this._disableDropdown;
+  }
   set disableDropdown(disableDropdown) {
     this._disableDropdown = coerceBooleanProperty(disableDropdown);
   }
 
   @Input()
-  get closeOnSelect() { return this._closeOnSelect; }
+  get closeOnSelect() {
+    return this._closeOnSelect;
+  }
   set closeOnSelect(closeOnSelect) {
     this._closeOnSelect = coerceBooleanProperty(closeOnSelect);
   }
 
   @Input()
-  get closeOnBodyClick() { return this._closeOnBodyClick; };
+  get closeOnBodyClick() {
+    return this._closeOnBodyClick;
+  }
   set closeOnBodyClick(closeOnBodyClick) {
     this._closeOnBodyClick = coerceBooleanProperty(closeOnBodyClick);
   }
 
   @Input()
-  get filterable() { return this._filterable; }
+  get filterable() {
+    return this._filterable;
+  }
   set filterable(filterable) {
     this._filterable = coerceBooleanProperty(filterable);
   }
 
   @Input()
-  get required() { return this._required; }
+  get required() {
+    return this._required;
+  }
   set required(required) {
     this._required = coerceBooleanProperty(required);
   }
 
   @Input()
-  get filterCaseSensitive() { return this._filterCaseSensitive; }
+  get filterCaseSensitive() {
+    return this._filterCaseSensitive;
+  }
   set filterCaseSensitive(filterCaseSensitive) {
     this._filterCaseSensitive = coerceBooleanProperty(filterCaseSensitive);
   }
 
   @Input()
-  get tagging() { return this._tagging; }
+  get tagging() {
+    return this._tagging;
+  }
   set tagging(tagging) {
     this._tagging = coerceBooleanProperty(tagging);
   }
 
   @Input()
-  get multiple() { return this._multiple; }
+  get multiple() {
+    return this._multiple;
+  }
   set multiple(multiple) {
     this._multiple = coerceBooleanProperty(multiple);
   }
 
   @Input()
-  get disabled() { return this._disabled; }
+  get disabled() {
+    return this._disabled;
+  }
   set disabled(disabled) {
     this._disabled = coerceBooleanProperty(disabled);
   }
@@ -161,7 +189,9 @@ export class SelectComponent implements ControlValueAccessor, OnDestroy {
   readonly inputComponent: SelectInputComponent;
 
   @ContentChildren(SelectOptionDirective)
-  get optionTemplates() { return this._optionTemplates; }
+  get optionTemplates() {
+    return this._optionTemplates;
+  }
   set optionTemplates(val: QueryList<SelectOptionDirective>) {
     this._optionTemplates = val;
 
@@ -203,7 +233,9 @@ export class SelectComponent implements ControlValueAccessor, OnDestroy {
     return this.placeholder && this.placeholder.length > 0;
   }
 
-  get value() { return this._value; }
+  get value() {
+    return this._value;
+  }
   set value(val: any[]) {
     if (val !== this._value) {
       this._value = val;
@@ -247,7 +279,7 @@ export class SelectComponent implements ControlValueAccessor, OnDestroy {
     private readonly _element: ElementRef,
     private readonly _renderer: Renderer2,
     private readonly _cdr: ChangeDetectorRef
-  ) { }
+  ) {}
 
   ngOnDestroy(): void {
     this.toggleDropdown(false);
@@ -333,8 +365,8 @@ export class SelectComponent implements ControlValueAccessor, OnDestroy {
     this._cdr.markForCheck();
   }
 
-  onKeyUp({ event, value }: { event: KeyboardEvent; value?: string; }): void {
-    if (event && event.key === KeyboardKeys.ARROW_DOWN as any) {
+  onKeyUp({ event, value }: { event: KeyboardEvent; value?: string }): void {
+    if (event && event.key === (KeyboardKeys.ARROW_DOWN as any)) {
       ++this.focusIndex;
     } else {
       this.filterQuery = value;
@@ -370,10 +402,10 @@ export class SelectComponent implements ControlValueAccessor, OnDestroy {
   /* istanbul ignore next */
   private onTouchedCallback() {
     // placeholder
-  };
+  }
 
   /* istanbul ignore next */
   private onChangeCallback(_: any) {
     // placeholder
-  };
+  }
 }