+++ projects/swimlane/ngx-ui/src/lib/pipes/index.ts
@@ -2,3 +2,4 @@ export * from './pipes.module';
 export * from './filter.pipe';
 export * from './decamelize.pipe';
 export * from './timezone.pipe';
+export * from './json-tree.pipe';